Single feature node regression

+++ realworld_benchmark/data/molecules.py
@@ -62,7 +62,7 @@ class StructureAwareGraph(torch.utils.data.Dataset):
             edge_list = (adj != 0).nonzero()  # converting adj matrix to edge_list
 
             edge_idxs_in_adj = edge_list.split(1, dim=1)
-            edge_features = adj[edge_idxs_in_adj].reshape(-1).float()
+            edge_features = adj[edge_idxs_in_adj].reshape(-1).long()
 
             # Create the DGL Graph
             g = dgl.DGLGraph()
